+++ kroxylicious-kms-providers/kroxylicious-kms-provider-azure-key-vault-kms/src/main/java/io/kroxylicious/kms/provider/azure/config/AzureKeyVaultConfig.java
@@ -17,8 +17,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import io.kroxylicious.kms.provider.azure.config.auth.EntraIdentityConfig;
-import io.kroxylicious.kms.provider.azure.config.auth.ManagedIdentityConfig;
+import io.kroxylicious.kms.provider.azure.config.auth.ManagedIdentityCredentialsConfig;
+import io.kroxylicious.kms.provider.azure.config.auth.Oauth2ClientCredentialsConfig;
 import io.kroxylicious.proxy.config.tls.Tls;
 import io.kroxylicious.proxy.tag.VisibleForTesting;
 
@@ -27,17 +27,18 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 /**
- * @param entraIdentity optional credentials for authenticating with Entra. Exactly one of either {@code entraIdentity} or {@code managedIdentity} must be specified.
- * @param managedIdentity optional service details for authenticating with Managed Identity. Exactly one of either {@code entraIdentity} or {@code managedIdentity} must be specified.
+ * @param oauth2ClientCredentials optional credentials for authenticating with Entra via OAuth2. Exactly one of either {@code oauth2ClientCredentials} or {@code managedIdentityCredentials} must be specified.
+ * @param managedIdentityCredentials optional service details for authenticating with Entra via Azure Managed Identity. Exactly one of either {@code oauth2ClientCredentials} or {@code managedIdentityCredentials} must be specified.
  * @param keyVaultName required name of the key vault to use for encryption, e.g. my-key-vault
  * @param keyVaultHost required host of key vault (without key vault name) e.g. vault.azure.net
  * @param keyVaultScheme optional scheme for making HTTP requests to key vault, default value is 'https'
  * @param keyVaultPort optional port for key vault (typically would only be used for testing), defaults to null implying no port will be included in requests
  * @param tls optional TLS configuration for key vault requests
  */
-@JsonPropertyOrder({ "entraIdentity", "managedIdentity", "keyVaultScheme", "keyVaultName", "keyVaultHost", "keyVaultPort", "omitVaultNameFromHost", "tls" })
-public record AzureKeyVaultConfig(@JsonInclude(NON_NULL) @Nullable @JsonProperty EntraIdentityConfig entraIdentity,
-                                  @JsonInclude(NON_NULL) @Nullable @JsonProperty ManagedIdentityConfig managedIdentity,
+@JsonPropertyOrder({ "oauth2ClientCredentials", "managedIdentityCredentials", "keyVaultScheme", "keyVaultName", "keyVaultHost", "keyVaultPort", "omitVaultNameFromHost",
+        "tls" })
+public record AzureKeyVaultConfig(@JsonInclude(NON_NULL) @Nullable @JsonProperty Oauth2ClientCredentialsConfig oauth2ClientCredentials,
+                                  @JsonInclude(NON_NULL) @Nullable @JsonProperty ManagedIdentityCredentialsConfig managedIdentityCredentials,
                                   @JsonProperty(required = true) String keyVaultName,
                                   @JsonProperty(required = true) String keyVaultHost,
                                   @JsonInclude(NON_NULL) @Nullable @JsonProperty String keyVaultScheme,
@@ -50,11 +51,12 @@ public record AzureKeyVaultConfig(@JsonInclude(NON_NULL) @Nullable @JsonProperty
     public AzureKeyVaultConfig {
         Objects.requireNonNull(keyVaultName);
         Objects.requireNonNull(keyVaultHost);
-        if (entraIdentity != null && managedIdentity != null) {
-            throw new IllegalArgumentException("more than one authentication method specified, must configure exactly one of either entraIdentity or managedIdentity");
+        if (oauth2ClientCredentials != null && managedIdentityCredentials != null) {
+            throw new IllegalArgumentException(
+                    "more than one authentication method specified, must configure exactly one of either oauth2ClientCredentials or managedIdentityCredentials");
         }
-        else if (entraIdentity == null && managedIdentity == null) {
-            throw new IllegalArgumentException("no authentication specified, must configure exactly one of either entraIdentity or managedIdentity");
+        else if (oauth2ClientCredentials == null && managedIdentityCredentials == null) {
+            throw new IllegalArgumentException("no authentication specified, must configure exactly one of either oauth2ClientCredentials or managedIdentityCredentials");
         }
         String host = URI.create("https://" + keyVaultHost).getHost();
         if (!Objects.equals(host, keyVaultHost)) {